backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pandas==2.1.3
openai==1.3.0
python-dotenv==1.0.0
pydantic==2.4.2
numpy==1.26.2
openpyxl==3.1.2
matplotlib==3.8.2
seaborn==0.13.0
httpx==0.26.0
aiofiles==23.2.1 